refactor: extract session header
