+++ catalyst_voices/uikit_example/lib/examples/voices_spaces_example.dart
@@ -1,11 +1,9 @@
 import 'package:catalyst_voices/widgets/app_bar/voices_app_bar.dart';
 import 'package:catalyst_voices/widgets/drawer/voices_drawer.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
-import 'package:catalyst_voices_assets/catalyst_voices_assets.dart';
-import 'package:catalyst_voices_blocs/catalyst_voices_blocs.dart';
 import 'package:catalyst_voices_brands/catalyst_voices_brands.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:uikit_example/widgets/session_header.dart';
 
 class VoicesSpacesExample extends StatelessWidget {
   static const String route = '/spaces-example';
@@ -19,7 +17,7 @@ class VoicesSpacesExample extends StatelessWidget {
     return Scaffold(
       appBar: const VoicesAppBar(
         actions: [
-          _SessionHeader(),
+          SessionHeader(),
         ],
       ),
       drawer: const VoicesDrawer(children: []),
@@ -52,88 +50,6 @@ class VoicesSpacesExample extends StatelessWidget {
           ),
         ],
       ),
-    );
-  }
-}
-
-class _SessionHeader extends StatelessWidget {
-  const _SessionHeader();
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<SessionBloc, SessionState>(
-      builder: (context, state) {
-        return switch (state) {
-          VisitorSessionState() => const _VisitorSessionHeader(),
-          GuestSessionState() => const _GuestSessionHeader(),
-          ActiveUserSessionState() => const _ActiveUserSessionHeader(),
-        };
-      },
-    );
-  }
-}
-
-class _VisitorSessionHeader extends StatelessWidget {
-  const _VisitorSessionHeader();
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        VoicesFilledButton(
-          child: Text('Get Started'),
-          onTap: () =>
-              context.read<SessionBloc>().add(const NextStateSessionEvent()),
-        ),
-        const SizedBox(width: 12),
-        VoicesFilledButton(
-          child: Text('Visitor'),
-        ),
-      ],
-    );
-  }
-}
-
-class _GuestSessionHeader extends StatelessWidget {
-  const _GuestSessionHeader();
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        VoicesFilledButton(
-          trailing: Icon(CatalystVoicesIcons.lock_open),
-          child: Text('Unlock'),
-          onTap: () =>
-              context.read<SessionBloc>().add(const NextStateSessionEvent()),
-        ),
-        const SizedBox(width: 12),
-        VoicesFilledButton(
-          child: Text('Guest'),
-        ),
-      ],
-    );
-  }
-}
-
-class _ActiveUserSessionHeader extends StatelessWidget {
-  const _ActiveUserSessionHeader();
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        VoicesFilledButton(
-          trailing: Icon(CatalystVoicesIcons.lock_closed),
-          child: Text('Lock'),
-          onTap: () =>
-              context.read<SessionBloc>().add(const NextStateSessionEvent()),
-        ),
-        const SizedBox(width: 12),
-        VoicesFilledButton(
-          child: Text('John Smith'),
-        ),
-      ],
     );
   }
 }
